# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TeamKnowledge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

  # ── Skills ────────────────────────────────────────────────────────────────
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List all skills
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Skills list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Create a new skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{skillId}:
    get:
      operationId: getSkill
      tags: [skills]
      summary: Get a skill with users who have it
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Skill detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillDetail"
        "404":
          description: Not found
    patch:
      operationId: updateSkill
      tags: [skills]
      summary: Update a skill
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete a skill
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── User Skills (associações pessoa ↔ skill) ──────────────────────────────
  /users/me/skills:
    get:
      operationId: listMySkills
      tags: [skills]
      summary: List current user skills
      responses:
        "200":
          description: User skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSkill"
    post:
      operationId: addMySkill
      tags: [skills]
      summary: Add a skill to current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSkillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSkill"
  /users/me/skills/{skillId}:
    patch:
      operationId: updateMySkill
      tags: [skills]
      summary: Update proficiency level for a skill
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSkillUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSkill"
    delete:
      operationId: removeMySkill
      tags: [skills]
      summary: Remove a skill from current user
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /users/{userId}/skills:
    get:
      operationId: listUserSkills
      tags: [skills]
      summary: List skills of a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSkill"

  # ── Knowledge (articles + curated links) ──────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, name, email, createdAt]
      properties:
        id:            { type: integer }
        clerkId:       { type: string }
        name:          { type: string }
        email:         { type: string }
        avatarUrl:     { type: ["string", "null"] }
        bio:           { type: ["string", "null"] }
        area:          { type: ["string", "null"] }
        createdAt:     { type: string }
    UserUpdate:
      type: object
      properties:
        name:      { type: string }
        bio:       { type: string }
        area:      { type: string }
        avatarUrl: { type: string }

    # ── Skill ─────────────────────────────────────────────────────────────────
    Skill:
      type: object
      required: [id, name, tags, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        usersCount:  { type: integer }
        createdAt:   { type: string }
    SkillDetail:
      type: object
      required: [id, name, tags, createdAt, users]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        usersCount:  { type: integer }
        createdAt:   { type: string }
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserSkill"
    SkillInput:
      type: object
      required: [name]
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
        tags:
          type: array
          items: { type: string }
    SkillUpdate:
      type: object
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
        tags:
          type: array
          items: { type: string }
    UserSkill:
      type: object
      required: [id, userId, skillId, level, skill, createdAt]
      properties:
        id:        { type: integer }
        userId:    { type: integer }
        skillId:   { type: integer }
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        skill:
          $ref: "#/components/schemas/Skill"
        user:
          $ref: "#/components/schemas/User"
        createdAt: { type: string }
    UserSkillInput:
      type: object
      required: [skillId, level]
      properties:
        skillId: { type: integer }
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
    UserSkillUpdate:
      type: object
      required: [level]
      properties:
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]

    # ── Knowledge ─────────────────────────────────────────────────────────────
...