# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeamKnowledge API
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  tags: string[];
  usersCount?: number;
  createdAt: string;
}