# lib/api-zod/src/generated/types/skillUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeamKnowledge API
 * OpenAPI spec version: 0.1.0
 */

export interface SkillUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  tags?: string[];
}