# lib/api-zod/src/generated/types/userSkillInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeamKnowledge API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSkillInputLevel } from './userSkillInputLevel';

export interface UserSkillInput {
  skillId: number;
  level: UserSkillInputLevel;
}