# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeamKnowledge API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  bio?: string;
  area?: string;
  avatarUrl?: string;
}